Extract output directory formatter, add tests, consolidate redundant pcap filter unit test classes.
